More Fixes and Added More Options
I originally implemented our game's inventory incorrectly. Previously,
I had an inventory for a plate and ingredient, but we changed our game
to only have one inventory that can take in a plate or ingredient. Lots
of interactions had to be changed for this implementation. I also added
other options like being able to inspect the current recipe and also
the ability to trash items. This trashing ability helped get rid of the
bottleneck when a person's inventory and the station ingredient were
both full which would not allow a person to do anything at that station.
I also added a recipe class (Target_Recipes.py) that would hold all the
recipes and possible ingredients; whenever we want to add a new
ingredient or recipe, we can just add it to that file and the game
should automatically incorporate the new additions. Last of all, I made
the game randomly choose a recipe and whenever a player submits their
plate, they will get a new recipe (or the same recipe if they happen to
randomly get the same recipe). A basic score keeps track of how many
recipes a player successfully submitted.