Improved Speech Detection
Added the pronouncing library and now our speech detection will check if
the word the speech API picks up is either: 1) the exact word, 2) rhymes
with the word we want, 3) the first 2 phonetic sounds of the word
matches the first 2 phoentic sounds of the intended word. If either of
these are satisfied will be considered valid.