More UI Changes; Fixed Multiplayer and Color
Added more UI changes, had to fix some multiplayer aspects that were not
working to fit with the UI. Also made the color tracking disabled when a
button is pressed. Some new issues is that speech is not translating to
animations, the player model dissapeared, and need a way to disconnect
when player enter's their name to play if they press the back button.
Problem is that if the player presses back and tries to go back on the
server, the server thinks the user is already on from the prev attempt.