Added Files For SinglePlayer, More UI
Added old singleplayer files for the sake of running main.py. Will spend
some time later condensing the code down so we have one file of each
that can run on multiplayer and singleplayer. Also added UI that shows
what voice commands are valid depending on where the player is. Yet, I
did not include the counters at the top of the kitchen since it is a bit
weird (for example, bottom of fridge is considred a bottom counter).
Thus, I recommend perhaps changing the color of valid counters so
players will always avoid trying to place items on counters at the top
of the kitchen.